+++ src/test/java/UserServiceTest.java
@@ -10,8 +10,7 @@ public class UserServiceTest {
     private final UserService userService = new UserServiceImpl();
     private final String testName = "Ivan";
     private final String testLastName = "Ivanov";
-    private final byte testAge = 5;
-
+    private final int testAge = 5;
 
     @Test
     public void dropUsersTable() {
@@ -60,7 +59,7 @@ public class UserServiceTest {
             userService.dropUsersTable();
             userService.createUsersTable();
             userService.saveUser(testName, testLastName, testAge);
-            userService.removeUserById(1L);
+            userService.removeUserById(1);
         } catch (Exception e) {
             Assert.fail("При тестировании удаления пользователя по id произошло исключение\n" + e);
         }
